+++ processor/src/test/resources/bad_input/com/example/component/BadTypeComponent.java
@@ -1,0 +1,8 @@
+package com.example.component;
+
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent( type = "" )
+public class BadTypeComponent
+{
+}